+(*) u,v tendendy diagnostic fix
Recent testing showed a difference in the time derivative diagnostics
dudt, dvdt, dhdt for symmetric and non-symmetric grids.  The issue
appears to be be due to assumed 1-indexing within the
`register_time_deriv` functions of MOM_diagnostics, due to use of
pointers without explicit start index.  This was causing an error in
`calculate_derivs`, which assume standard grid indexing.

There was no error in non-symmetric mode, which assumes 1-indexing (when
global indexing is disabled, as default), but was creating off-by-one
errors in symmetric mode for the u and v time derivaties, which
typically start at 0 for the x and y axes, respectively.  The dhdt
diagnostic was unaffected, since h-points always use 1-indexing.

This also affected the dKE/dt diagnostic, since they used dudt and dvdt.

We resolve this by introducing a lower bound input argument to
`register_time_deriv` and use this to define the start index for the
various p3d pointers.

This is a minor API change which only affects the dudt, dvdt, and dhdt
diagnostics, and may change the dudt and dvdt symmetric grid diagnostics.