(*) speed diagnostic fix on nonsymmetric grids
The speed diagnostic currently requires an interpolation of u and v onto
the h-points.  On non-symmetric grids, this was causing an error on the
west and south boundaries, since the compute loops in
`extract_surface_state` did not update the boundaries, which are used in
the speed calculation.

We resolve this by extending the loops from G%IscB, which only includes
the boundary on symmetric grids, to isc-1, which always include the
boundaries.  This assumes a nonzero halo, and that the halos of u and v
have been updated.