dKEdt diagnostic nonsymmetric bugfix
This patch fixes a minor error in the d(KE)/dt diagnostic, specifically
any values on the western and southern boundaries (per PE domain).

d(KE)/dt depends on the dudt and dvdt diagnostics.  However, the western
and southern boundaries of these two diagnostics were not updated inside
`calculate_derivs`, which caused an error on the western and southern
boundaries of d(KE)/dt since it requires an interpolation of dudt and
dvdt onto h-points.

This was not an issue in symmetric grids, since the west/south values
would have been included in the update, but were not updated in
non-symmetric mode.

We resolve this by extending the loops to update an additional grid
point to the west/south.  This slightly increases the computation, and
assumes an updated halo, but this change has restored consistency of
d(KE)/dt in symmetric and nonsymmetric mode.