(*) Vertical diagnostic remap symmetric bugfix
This patch fixes an error in the vertial diagnostic remapping on
symmetric grids.

The following functions associated with vertical remapping assume
grid-based indexing, but receive arrays with 1-based indexing:

* diag_remap_do_remap
* vertically_reintegrate_diag_field
* vertically_interpolate_diag_field

Since the calculations typically include a mix of grid-based and 1-based
indexing, we retain the grid index loops but introduce 1-based indices,
i1 and j1, when addressing these arrays.

Note that while the symmetric grid u- and v-points are now defined
relative to the global grid (G%isg, G%jsg), we still assume that the
h-points are locally 1-based.  This is currently true in MOM6 but may
require modification if we move to global grid indexing.

We have included a very verbose explanation for the gridding issue, in
case it ever gets addressed in the future.

Symmetric grid results now match non-symmetric grid results, and
non-symmetric grid results are unchanged.  Model state and dynamics are
unaffected by this change and should be bitwise reproducible.