Code clean-up in MOM_energetic_PBL
  Various code clean-up in MOM_energetic_PBL, including the following: Corrected
the indenting.  Commented out some variables that were only of use for debugging
during model development.  Restored several lines that do simple assignments on
a single line with semicolons.  Followed MOM6 standards with regard to the
non-use of argument names for non-optional arguments.  Corrected the case of the
K index for the interface variable Kd in a number of places. Added some comments
documenting various options.  Added units to the comments describing several
subroutine arguments.  All answers are bitwise identical and there are no
changes to MOM_parameter_doc files.