(*) Remapped d[uv]dt_dia, [uv]hml diagnostic fixes
The remapped diagnostics d[uv]dt_dia, and uhml were reporting different
results for different CPU layouts.  This was due to a dependence on the
east and north halo values of the h and remap_cs%h thicknesses, which
were not being updated for these variables.

The remap_cs%h thickness directly depends on h, and will properly fill
in its halo values, but is only updated when a `diag_update_remap_grids`
function is called.  If h has missing halo values prior to this call,
then remap_cs%h will also be missing.

In most of the code, h has been properly updated, and remap_cs%h only
needs an update when h changes, which is restricted to a few functions.
This appears to not be an issue for most diagnostics, with the ones
above being an exception.

We resolve this by applying a halo update to h prior to any
`diag_update_remap_grids` calls which are needed before the diagnostic
`post_data` call.  We only apply this halo update if the diagnostic has
been registered.

Ideally we would only apply the halo update when the remapped halo has
been registered, but it is currently not possible to query this
information, and ought to be done in a future release.

It is likely that there are more diagnostics with this issue, so
consider this a first step.