(*) Z-remap interface mask{Cu,Cv,B}i bugfix
Z-interpolated diagnostics on interfaces were failing to produce
consistent answers when CPU layout was changed.  This was tracked down
to an error in the eastern boundaries of their mask fields.

These mask fields are constructed from the T-point mask, maskTi, whose
halo values were not being constructed.  By extending the loops from
[ij]sc..[ij]ec to [ij]sc-1 .. [ij]ec+1, we restored the masks and could
reproduce the values.

This will affect the following diagnostics when z-remapped:

* neutral_slope_[xy]
* KHTH_[uv]

and potentially any other diagnostics on the `axesCui` axis.