(*) Horizontal remapping re-enabled, bugfixes
This patch re-enabled the `post_xy_average` calls in `post_data_3d_low`,
which are used to generate and write the horizontally averaged
diagnostics.

We have also fixed a number of issues in the existing methods, and we
have changed how masked values are handled.

The first major change is that average of a field is now computed on
the native grid points, rather than interpolation onto h-points.  This
is due to a dependence on halo values of `h` which would be potentially
undefined in symmetric grids.  In most cases the calculation will be
mathematically equivalent, although there will be bitwise differences.
Non-periodic boundaries may also yield different edge values which will
modify the average.

The second major change is the introduction of a reproducible sum for
the calculation of the total volume and field at each vertical level,
which will introduce bitwise changes but is now reproducible across
layouts.

The final change is the removal of explicit `missing_value` checks,
which have been replaced with a new column mask.  Masked values are now
explicitly set to zero.  This simplifies the calculation of any averages
or checksums, and is also consistent with the overall removal of
explicit missing_value usage in MOM.

These changes have removed depenendies on unassigned values in symmetric
grids, and produce results which are now reproducible across layouts and
for symmetric and nonsymmetric grids.  Diagnostic values are expected to
change, but the model state should be bitwise reproducible.