Do not attach cell methods in diag chksum mode
The diag mediator register functions were using the `attach_cell_method`
and `diag_field_add_attribute` functions to attach attributes to the FMS
diagnostics.  This required referencing of internal FMS field tables,
which are not created when DIAG_AS_CHKSUM is enabled.

While this did not raise any errors in the Baltic tests (for unclear
reasons), this was causing issues in nearly every other test,
predominantly when not in debug mode.

We now only conditionally call these functions when `diag_as_chksum` is
disabled.