Repro sum increment_ints_faster overflow check
While reproducing_sum was able to detect an overflow for values
exceeding its maximum allow value, `prec**N_max * (2**63 - 1)`, this
check occurs after the first bin decomposition, which requires an
integer conversion, `int(r * (1./prec_max), 8)`.

When floating point exceptions (FPE) were disabled, this would often
produce a large negative integer, and the overflow would have been
caught later and handled as prescribed by the user.  But when FPEs are
enabled, there is a fatal error and the model aborts.  (NOTE: The FPE
handler in GCC tags this as "invalid", rather than "overflow")

This patch adds a check inside `increment_ints_faster`, so that the
`int()` operation is avoided in the case of overflow.