Flux allocation check in diagnostics
Currently the model will attempt to compute any enabled surface flux
diagnosics, even if the model has not been configured to support this
flux.  This would raise memory access errors if the flux had not been
allocated.

We now conditionally check if a particular surface flux has been
allocated before adding it to any cumulative flux diagnostic.

In the future, we should probably implement some framework around all
this and associate the diagnostic with the creation of the flux.