Conditionally enable MEKE Kh diagnostics
It is possible to enable the KhMEKE_u and KhMEKE_v diagnostics when
their values have not been updated, which will either write nonsense
values or raise errors/exceptions if checksums are enabled.

This patch introduces a general flag to check if lateral
(non-biharmonic) MEKE fluxes have been turned on, and only registers the
KhMEKE_[uv] variables if they are present.