+(*)Corrected documentation of BBL_MIXING_MAX_DECAY
  Revised the get_param message for BBL_MIXING_MAX_DECAY to be consistent with
how the code actually works, and changed the code to handle 0 values as
documented. Also altered the default value to 200 m so that the answers will
reproduce the previous solutions, unless BBL_MIXING_MAX_DECAY was explicitly
being set to 0 or a negative value, in which case the revised code will match
the intended behavior as documented.  All answers are bitwise identical in the
existing MOM6-examples, but some MOM_parameter_doc files change and some
solutions could change, in which case setting BBL_MIXING_MAX_DECAY=200 will
reproduce the previous solutions.