+Added WIND_GYRES_2018_ANSWERS
  Added the new runtime parameter WIND_GYRES_2018_ANSWERS to enable the
transition to newer and simpler expressions for ustar in the gyres option of the
solo_driver version of MOM_surface_forcing.  Also replaced the markers in the
comments around the controlled-forcing code with #CTRL# to distinguish them from
other comments.  By default all answers are bitwise identical, but there is a
new entry in some MOM_parameter_doc files.