(*)Rescaled velocities in MOM_kappa_shear to L T-1
  Rescaled the horizontal velocities in MOM_kappa_shear to L T-1 for dimensional
consistency testing.  This does not change answers after dimensional rescaling
provided that VEL_UNDERFLOW is a small positive number (like 1e-30 m s-1), and
in some cases even if VEL_UNDERFLOW is 0, and it does not change answers when
there is no dimensional rescaling for length or time. All answers are bitwise
identical in the MOM6-examples test cases.