(*) Stronger conditional register of FrictWorkMax
The MEKE criteria for using the `FrictWorkMax` diagnostic was not
sufficient, since it is possible to use MEKE while still unable to
compute some of the terms required for `FrictWorkMax`.

We resolve this by adding the `MEKE_type` struct as an argument for
`hor_visc_init`, and use the same information when computing
`FrictWorkMax` to determine whether to register `FrictWorkMax`.

This changes the API by adding MEKE to most of the timestep
initializations, but should not affect answers.