+Added HOR_VISC_2018_ANSWERS & rescaled variables
  Added the runtime parameter HOR_VISC_2018_ANSWERS to permit the elimination of
a dimensional constant without changing answers. Rescaled the units of several
time variables in MOM_hor_visc.  Also added comments indicating issues with the
GME options and suggests for how to correct them.  All answers are bitwise
identical, but there is a new entry in the MOM_parameter_doc files.