Changed the timestep units to [T} in hor_visc_init
  Changed the units of the timestep to [T} in hor_visc_init and of vert_vort_mag
to [T-1 m-1] in horizontal_viscosity.  Also added a variant of the
grad_vel_mag_h calculation with parentheses for rotational symmetry when
answers_2018 = False.  Changed the marks around suggestions for correcting
issues with the recently added GME code to #GME# to help in finding them.  All
answers are bitwise identical.