+Changed the units of MEKE%Ku [m2 T-1]
  Changed the units of MEKE%Ku and MEKE%Au to [m2 T-1], including adding code
to allow for the dimensional scaling to change across restarts and moving the
halo updates on any MEKE variables read from restart files to the end of
MEKE_init.  Also change the units of GME_coeff in horizontal_viscosity to
[m2 T-1].  This also required adding a unit_scale_type argument to MEKE_init.
All answers are bitwise identical, but the units for some variables in a
publicly visible type have changed.