+Rescaled elements of the BT_cont_type
  Added dimensional rescaling of the face-area and velocity elements of the
BT_cont_type for dimensional consistency testing.  This required the addition
of unit_scale_type arguments to several of the continuity-related subroutines.
All answers are bitwise identical, but there were changes to the units of
elements of a public type and there are new required arguments to several
public subroutines.