+Rescaled uhtr and vhtr
  Rescaled the units for the mass fluxes that do tracer transport to [L2 H] for
expanded dimensional consistency testing.  This required the addition of
unit_scale_type arguments to several routines.  In addition some commments were
added or corrected.  All answers are bitwise identical, but there are some minor
changes to public interfaces.  The offline tracer capability is not being tested
in the MOM6-examples test suite, so there is not guarantee that it will pass
the dimensional consistency tests, but the cases in the test suite all pass.