+Rescaled continuity internal calculations
  Applied dimensional rescaling to all of the velocities used inside of
continuity_PPM or passed to continuity_PPM from continuity to work in units of
[L T-1].  Also rearranged the dimensional scaling factors that will align with
the grid spacing to facilitate later cancellations.  This required the addition
of unit_scale_type arguments to the continuity initialization routines.  All
answers are bitwise identical, but the units of the arguments to a public
routine that is wrapped inside of another have changed, and there is a new
argument to publicly called routines.