Bugfix: diabatic_salt_tendency diagnostics in ALE
The diabatic_salt_tendency and diabatic_salt_tendency_2d diagnostics
relied on a scratch array (work_3d) which is only computed if
diabatic_diff_saln_tendency is also enabled.  If it is not set, then the
array is filled with values from an older diagnostic (e.g. temperature
tendency).

We resolve this by computing the scratch array when any of the three
diagnostics are enabled.  Internal logic is used to determine whether or
not to post the data.

This resolves GitHub issue #968.  Thanks to Graeme MacGilchrist
(@gmacgilchrist) for reporting.