Rescaled internal variables in MOM_hor_visc.F90
  Rescaled multiple internal variables in MOM_hor_visc.F90 for more complete
dimensional consistency testing.  One dimensionally inconsistent expression
(i.e., a bug) was identified and marked, but the code has not been changed yet
so that the answers do not change. It is conceivable that underflow would be an
issue in some test cases with out an explicitly set underflow velocity, but all
answers in the MOM6-examples test csaes are bitwise identical.