(*)Eliminated a bug in horizontal viscosity work
  Commented out code that deliberately retained a bug, thereby eliminating the
dimensionally inconsistent expressions for the limits on the estimates of energy
extracted by the horizontal viscosity.  The decision to simply eliminate this
bug by commenting out code instead of adding a run-time bugfix parameter was
taken after consulting with Scott Bachman and Malte Jansen, whose simulations
seemed the most likely to be impacted.  All answers are bitwise identical in the
dev/gfdl MOM6-examples test cases, but other solutions could be changed.