+Pass velocities to btstep in [L T-1]
  Passed the velocity arguments to btstep and barotropic_init in
rescaled units of [L T-1].  As a temporary step, array syntax is being used to
rescale arguments directly in some subroutine calls, even though this practice
is strongly discouraged in the MOM6 code standards.  All answers are bitwise
identical, but the rescaled units of several arguments in a public interface
have changed.