Travis test suite expansion
This patch extends the current test suite to report code coverage and
includes the following tests:

* Regression testing (for a GitHub PR)
* Symmetric/nonsymmetric grids
* Layouts for 1 and 2 domains
* Restart resubmission
* DEBUG/REPRO builds
* NaN initialization of allocated arrays
* Dimensional scaling (2^11)

The test uses four build configurations:

* Symmetric grids
* Asymmetric grids
* REPRO build
* A "target" build for PR merges, as a regression test

A test will pass if the "ocean.stats" file is unchanged when compared to
the symmetric DEBUG build.

Comparisons of the diagnostic checksums ("chksum_diag") are included
but the results are only reported and are not treated as a
fail-condition.

Four experiments are currently supported, currently based on the
MOM6-examples repo:

* tc0: unit_tests
* tc1: A low-resolution configuration of benchmark
* tc2: An ALE-based config of benchmark
* tc3: A shortened simulation of circle_obcs

Coverage reports are generated using "gcov" and are reported to
"codecov.io".

The build Makefile (.testing/Makefile) is usable at command line and
does not require a Travis VM session, although there are some
potentially volatile dependencies on git which should probably be
explored before distributing it for general use.

It is expected that the tests will be expanded (e.g. repeat using
MOM_parameter_doc output) as well as the number of experiments.

Other notes:
- Includes an independent build of FMS
- Real-number inequality warnings are disabled