Explicit dependency for Travis path_names Makefile
The path_names rules in the Travis Makefile were using a variable to
define its remote branch, which does not work in Makefiles.

This patch removes the REMOTE variable and the rule has been split into
two rules with explicit dependencies.  We might try to merge these two
rules at some point.

This issue was not raised in previous tests because there was an old
remote branch with the same name as the merged case (mergetest).  So
namespace conflicts are still a potential issue with this Makefile.