Makefile local support
The testing Makefile was updated to avoid aggressive git repo
manipulations when used locally.  When DO_REGRESSION_TESTS is disabled,
we now define the target and merge branches as the current local branch,
and most git operations become inert.

As part of this, the DO_REGRESSION_TESTS variable is handled as a more
traditional bash variable, where false is attributed to unassigned
values, rather than explicitly looking for the "false" string.