Skip checkout/merge for non-regression tests
Out Travis testing makefile relied on `git rev-parse` to determine its
branch name, but this does not work for pushes which returns the name of
HEAD.

We were running the non-regression builds through the same rule which
merges the target reference codebase, but were relying on these to be
null operations since they were operating on themselves.  But this was
always a poor strategy.

We resolve this by un-defining the dependency on the rule which creates
this new merged branch, and should skip the checkout/merge step
altogether.  (But we shall see...)