Travis: git management replaced with subdirectory
The manipulation of the repository in-place has been replaced with a
self-checkout of the target codebase inside of the build/ directory of
the repository, which is used to build the target executable.

This eliminates the need for self-manipulation of the repository during
testing, as well as much of the complexity of the existing Makefile.