Travis: OpenMPI environment variable support
Travis OpenMPI jobs appears to not be supporting the
MALLOC_PERTURB_ environment variable, most likely because we are not
using the -x flag.

This PR introduces a $(MPIRUN_CMD) Make function to support more
generalized mpirun commands.  We can expand it in the future to include,
say, the MPI rank count.