Initialize geothermal internal heat diagnostics
The variables used for calculating internal heat in MOM_geothermal.F90
relied on zero-initialization due to the conditional do_i() check inside
of the loops, which was producing NaN warnings in our test suite.

This patch resolves this by conditionally initializing the fields based
on diagnostic registration.

It also introduces two new logical variables to control the
initialization and evaluation of the arrays.