PRECIOUS output; gitignore restructure
ocean.stats and chksum_diag output is now tagged as PRECIOUS, and is
retained after a run.  This helps to identify an issue after a fail.

.gitignore was also restructured to move the test ignore rules into a
subdirectory.

Finally, we no longer rely on (-) to ignore commands in Makefile, since
it was causing odd issues after PRECIOUS was introduced.  We now just
explicitly ignore any commands by piping nonzero commands to `true`.