(*) Bugfix: [uv]hml z-diags in general restrat
The z-interpolated uhml and vhml diagnostics gave inconsistent answers
across layouts when using the general mixed layer restratification
scheme, because the value of h had changed but its halos had not been
updated.

This had been previously fixed in the BML restratification but not the
general stratification method.

This patch updates the value of h by conditionally updating the halos if
this diagnostic is required.