(+) Tracer diffusive flux diagnostic scaling
Diagnostics of diffusive tracer fluxes (*_diff[xy], *_diff[xy]_2d) were
not scaling due to issues with the Coef_x term, which was in units of
L2 while the tracer itself was in units of m2.

Although the Coef_x term was being de-scaled to meters, the cumulative
sum was failing to reproduce for many tracers, either being reverted (in
temp) or failing to reproduce all bits (in salt).

We resolve this by re-defining the diffusive flux terms as H L2 s-1,
rather than H m2 s-1, and moving the scaling to the final conversion
factor, which restored reproducibility of all tracers.

This patch contains an API change.  We have added the unit scaling
struct (US) to the register_tracer_diagnostic function.