Unsplit timestep diag fixes (CAu, etc)
This patch fixes a bug in the registration of CAu, where the conversion
argument had been included in the string of the dimensions of the
diagnostic registration.

The units of various accelerations have also been changed from 'meter
second-2' to 'm s-2' for consistent with the rest of the model.