Add function and unit tests for boundary layer mixing
For the near-boundary lateral mixing, the indices of the layers that
are spanned by the boundary layer need to be returned. Additionally,
in cases where the boundary layer intersects partway through a layer,
the non-dimensional position also needs to be returned for polynomial
reconstructions to be evaluated correctly. Six unit tests were added
to test this new function.

All unit tests currently pass