(+) MKS thickness scaling of horizontal avg diag
The horizontal_average_diag_field function uses thickness when computing
layer averages in some cases.  Reproducibility requires that
reproducible sums be calculated in MKS units.  While the grids have been
converted to MKS lengths, the thickeness have not.

This patch rescales the thicknesses to volumes, which restores the
reproducibility of the dimentionally scaled quantities involving H.