Diagnostic dimensional scaling fixes
This fixes up the dimensional scaling of several diagnostics:

* dynamics_h_tendency
* MLD_Restrat
* ea, eb
* h_predia
* frazil_h

We also include a change to the flux_to_kg_per_s scaling factor.
Previously, this was dynamically set the flux units to either m3 s-1 or
kg s-1 depending on the Boussinesq mode.  However, the units of the
diagnostics using this factor (uhml, vhml) were always explicitly set to
kg -1.  Assuming that the scaled fluxes were H L2 T-1, this would have
given the wrong scaling in non-Boussinesq mode.

We resolve this by removing the conditional and always using
GV%H_to_kg_m2 when defining this scaling factor.