+Add density rescaling factors in unit_scale_type
  Added factors for power-of-2 rescaling of density to the unit_scale_type,
along with the new run-time parameter R_RESCALE_POWER.  All answers are
bitwise identical, but there is a new runtime parameter, some new elements in
a transparent public type, and a new optional variable in the MOM restart
files.  This adds a new entry to the MOM_parameter_doc.debugging files.