Update LBM flux names and fix posting of 2d diags
The 2d diagnostics associated with the lateral boundary mixing were
occuring outside the `if (CS%id_lbm_dfx_2d)` statements and were
thus leading to segfaults if the diagnostic was not requested.

Additionally the array variable names were refactored to be consistent
with the `id_` names and the diagnostic names.