Report negative zero min/max as positive
This patch addresses the inconsistency of signed zero in the minimum and
maximum values used in checksum report.

The behavior of the Fortran intrinsic min() and our MPI library's
implementation of MPI_Reduce with MPI_MIN can give different results for
different values of signed zero, e.g. min(0,-0) vs min(-0,0).
Additionally, the MPI_Reduce result appears to not consistenty follow
these rules in more complex MPI calculations.

Due to these issues, we add the result to positive zero to ensure that
any negative zero results are reported as positive.