Fix flux limiter in LBM when < 0
The previous flux limiter which doesn't allow the diffusive flux
to be greater than a 1/4 of the tracer inventory was incorrect
when the sign of the flux was negative. This has been fixed.

Incidentally, an additional 'if' condition was placed at the top
of the loop calculating the layer fluxes to avoid unnecessary
evaluations when we don't expect to be calculating a flux.