(*)+Rescaled MEKE source element units in MEKE_type
  Rescaled units of MEKE source elements in MEKE_type, and made slight code
modifications when MEKE_GM_SRC_ALT is true to ensure that the documented units
are always used.  Also corrected a bug in omitting looping over layers with the
same setting, which will change answers when MEKE_GM_SRC_ALT is true.  All
answers are bitwise identical in the MOM6-examples test cases.