Use popcnt intrinsic for bitcount
Profiling of the test suite showed a large amount of time (nearly 1/3)
devoted to computing of bitcounts used as checksums for diagnostics.

This patch replaces the bit loop with the popcnt intrinsic, which
produces the same result and uses the hardware assembly instruction when
available (e.g. popcnt in x86).

This change appears to have reduced the runtime of the test suite from
4.5 minutes to under 3 minutes.