+Rescaled the units of taux_shelf and tauy_shelf
  Rescaled the units of the taux_shelf and tauy_shelf elements of the
vertvisc_type to [R Z L T-2], and made these elements of the surface type into
allocatable arrays so that they can retain units of [Pa].  Also added code to
allocate these arrays as needed. In addition, commented out the improper code
setting taux_shelf and tauy_shelf as a non-vector in shelf_calc_flux, but this
was not being used anyway.  Also canceled out rescaling factors in the
expressions for taux_bot and tauy_bot in vertvisc.  All answers are bitwise
identical in the MOM6-examples test cases, and should be unaltered in other cases.