Adds MEKE_equilibrium_restoring
This commit adds a subroutine that calculates a new equilibrium value
for MEKE at each time step. This is not copied into MEKE%MEKE; rather,
it is used as a restoring term to nudge MEKE%MEKE back to an equilibrium value.
To select this option one needs to set MEKE_EQUILIBRIUM_RESTORING=True.
The timescale for nudging is controlled via MEKE_RESTORING_TIMESCALE.