Ice shelf and barotropic rescale variable init
Some index bounds in the ice shelf code were not initialized to the grid
values.  This patch sets those values.

The vel_rescale parameter used in barotropic_init was also absent from
one block of code, and has been added.  (This is probably a temporary
variable that will be phased out as dimensional scaling progresses.)

Presumably none of this code was being run, because it would have
normally caused severe problems.