Kh_max_x[xy], Ah_max_xy chksum and alloc fixes
This patch fixes the size of Kh_max_xx, which is on h-points but was
incorrectly allocated as if it were on B-points.  This also indirectly
fixes an issue with the calculation of its checksum.

This also fixes two checksum bugs, where the Kh_max_xy and Ah_max_xy
were incorrectly computing the Kh_max_xx and Ah_max_xx checksums
(respectively).

As a side note, this was incorrectly reporting an invariance of the
Kh_max_xy checksum, a B-point quanity, since it was computing the
Kh_max_xx checksum (h-point) on B-points.

This PR resolves the inconsistency of Kh_max_x[xy] and Ah_max_x[xy]
checksums across experiments, and slightly reduces Kh_max_xx to its
correct size.