Extend z_top and z_btm density calculation in diag
The calculate_vertical_integrals function in MOM_diagnostics includes
the calculation of density in the z_top and z_btm arrays.  The values
are initializes from is:ie and js:je, but the int_density_dz function
expects these arrays to be defined from isq:ieq+1 and jsq:jeq+1.

The reduced init range raises a FPE when signaling NaN initialization is
turned on.  Previously, these values would have presumably be set to
zero.

This patch extends the initialization of z_top and z_btm to prevent this
issue.