opacity: netPen explicit summation
The net penetrating shortwave heat flux at the top level was using colon
notation and a SUM() intrinsic, which was causing sums over NaN in
Nan-initialized mode, as well as a potentially ambiguous summation
ordering.

This patch replaces the implicit update and summation with explicit sums
over the reduced non-halo domain.