advect_x logical test split
A conditional test in advect_x of MOM_tracer_advect contained a logical
short-circuit:

    (do_i(i)) .and. (Ihnew(i) > 0.0)

where Ihnew was unset if do_i was false.  This caused an FPE in builds
where short-circuit logic checks were disabled and Ihnew was initialized
with NaN.

This patch fixes the issue by splitting the statement into two nested
if-statements.