MOM_entrain_diffusive logical nesting
Two logical checks in MOM_entrain_diffusive were restructured to
eliminate short-circuit logical checks which were raise when NaN
initialization was enabled.

In the first case, a new logical variable (do_entrain_eakb) was
introduced to enable nested-and conditions.

In the second case, a nested (k == kb) if-block inside of a (k >= kb)
block was moved outside as a separate if-block.