Testing: complete all regression tests
This patch modifies the test suite makefile to complete all regression
tests before reporting if there is a failure.  We achieve this by
storing the diff between the symmetric and target runs as separate
files; if any files exist (tested using 'ls') then we print the list of
files and report a failure.

This allows us to monitor all files which have changed, rather than just
the first run which has failed.

We do not yet extend this to the invariance tests, but may do so in the
future.