Travis: Move regression tests to separate build
This patch removes the regression test from the standard Travis job and
moves it to a new job.  We also run all regression tests before
reporting the status code, and report a fail if any of the test give a
different result.  All files which have changed are reported at the end
of the job.

Code coverage tests have also been moved out of the invariance tests and
into the regression testing.  They continue to be based on symmetric
grids.

This test also combines the whitespace and documentation auditing into a
a single test, in order to stay within 3 concurrent tests.